requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-HTTPAuth==3.2.4
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.3
passlib==1.7.1
PyMySQL==0.9.3
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.3
Werkzeug==0.15.2

alembic==0.8.9
bcrypt==3.1.2
cffi==1.9.1
coverage==4.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-Testing==0.6.1
Mako==1.0.6

pycparser==2.17
PyJWT==1.4.2
python-editor==1.0.3